move data_chunk_source_factories to source file
